quantify allele based on aligned bam
molid tags are in the read id, which can be used to quantify abundance
of alleles captured by a read. we also count mol id abundance by
allowing a number of mismatches, e.g. ndiff=2, which means two mol ids
with a number of differences less than ndiff would be considered as a
same ID.